+++ CHANGES.md
@@ -8,6 +8,7 @@ dev
 
 - Change `Pass:setScissor` to clamp negative x/y values to zero.
 - Change `lovr.math.gammaToLinear/linearToGamma` to return their 4th argument unchanged.
+- Change `Image:get/set/mapPixel` to support `rgb565`, `rgb5a1`, `rgb10a2` formats.
 
 ### Fix